Updated Docker configuration and added specific user
The Docker configuration has been updated to use version '3.8' and changes volumes configuration for a more managed data persistence. The Dockerfile has been updated to create a specific user, 'user_app', to run the app enhancing the security by avoiding running the app as a root user. The environment variables are also redefined to adjust to these changes.