Refactor UserRole import paths
In both user_dto.py and user_entity.py, the path from which the UserRole is imported was changed from 'constants.Enum.user_role' to 'nutri_app_core.constants.Enum.user_role'. This was done to reflect changes to the project's directory structure which were necessary for improving code maintainability and organization.