Implemented pagination for the User API routes
Added functionality for pagination to the 'get_all_users' API route. PaginationParams and PaginationResponseDto types from 'fastutils_hmarcuzzo' library were imported. These changes were made to improve the performance of the API by reducing the load time and overcoming the slow processing time when a large amount of data is fetched from the server all at once. In the future, the size of the page and the page number can be specified to only receive necessary data, reducing server load and improving user experience.